app/requirements.txt
PyYAML==6.0.2
requests==2.32.3
websocket-client==1.8.0
Flask==3.0.3
Flask-Cors==4.0.1
